[Translator] Deprecated transChoice and moved it away from contracts
