+++ src/Symfony/Component/Translation/Tests/TranslatorTest.php
@@ -393,6 +393,7 @@ class TranslatorTest extends TestCase
 
     /**
      * @dataProvider getTransChoiceTests
+     * @group legacy
      */
     public function testTransChoice($expected, $id, $translation, $number, $parameters, $locale, $domain)
     {
@@ -406,6 +407,7 @@ class TranslatorTest extends TestCase
     /**
      * @dataProvider      getInvalidLocalesTests
      * @expectedException \Symfony\Component\Translation\Exception\InvalidArgumentException
+     * @group legacy
      */
     public function testTransChoiceInvalidLocale($locale)
     {
@@ -418,6 +420,7 @@ class TranslatorTest extends TestCase
 
     /**
      * @dataProvider      getValidLocalesTests
+     * @group legacy
      */
     public function testTransChoiceValidLocale($locale)
     {
@@ -537,6 +540,9 @@ class TranslatorTest extends TestCase
         );
     }
 
+    /**
+     * @group legacy
+     */
     public function testTransChoiceFallback()
     {
         $translator = new Translator('ru');
@@ -547,6 +553,9 @@ class TranslatorTest extends TestCase
         $this->assertEquals('10 things', $translator->transChoice('some_message2', 10, array('%count%' => 10)));
     }
 
+    /**
+     * @group legacy
+     */
     public function testTransChoiceFallbackBis()
     {
         $translator = new Translator('ru');
@@ -557,6 +566,9 @@ class TranslatorTest extends TestCase
         $this->assertEquals('10 things', $translator->transChoice('some_message2', 10, array('%count%' => 10)));
     }
 
+    /**
+     * @group legacy
+     */
     public function testTransChoiceFallbackWithNoTranslation()
     {
         $translator = new Translator('ru');
